Consolidate CLI and config file configs
