+++ src/configs.cpp
@@ -30,7 +30,34 @@ fs::path get_proj_data_dir()
 
 } // namespace
 
-void Configs::load_configs_from_config_file()
+namespace params {
+
+void CommandLineParameters::validate_params()
+{
+    if (this->input_file.empty()) {
+        throw std::invalid_argument("No filename was provided. Cannot proceed");
+    }
+
+    if (this->instructions_file) {
+        if (this->instructions_file.value().empty()) {
+            throw std::invalid_argument("Instructions filename was not provided. Cannot proceed");
+        }
+    }
+
+    if (this->output_file) {
+        if (this->output_file.value().empty()) {
+            throw std::invalid_argument("Output filename was not provided. Cannot proceed");
+        }
+    }
+
+    if (this->instructions_from_cli) {
+        if (this->instructions_from_cli.value().empty()) {
+            throw std::invalid_argument("CLI instructions are empty. Cannot proceed");
+        }
+    }
+}
+
+void CommandLineParameters::load_configs_from_config_file()
 {
     static fs::path proj_config = get_proj_data_dir() / "funcgraft.toml";
 
@@ -57,4 +84,4 @@ void Configs::load_configs_from_config_file()
     }
 }
 
-Configs configs;
+} // namespace params
