+++ src/configs.hpp
@@ -1,9 +1,21 @@
 #pragma once
 
+#include <filesystem>
+#include <optional>
 #include <string>
 
-struct Configs {
+namespace params {
+
+struct CommandLineParameters {
     void load_configs_from_config_file();
+    void validate_params();
+
+    bool verbose = false;
+
+    std::filesystem::path input_file;
+    std::optional<std::filesystem::path> instructions_file;
+    std::optional<std::filesystem::path> output_file;
+    std::optional<std::string> instructions_from_cli;
 
     int port_ollama = 11434;
     std::string provider;
@@ -12,4 +24,4 @@ struct Configs {
     std::string model_openai;
 };
 
-extern Configs configs;
+} // namespace params
