+++ src/prompt/instructions.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "params.hpp"
+#include "configs.hpp"
 
 #include <string>
